General Refactoring for mobile app and update in directory links for webapp api
