+++ node_src/models/mongoose/donor.js
@@ -1,65 +1,58 @@
-const mongoose = require('mongoose');
-const bcrypt = require('bcrypt');
+const mongoose = require("mongoose");
+const bcrypt = require("bcrypt");
 
 //Creating a Donor Schema
-const donorSchema = new mongoose.Schema({
+const donorSchema = new mongoose.Schema(
+  {
     firstName: {
-        type: String,
-        required: true
+      type: String,
+      required: true,
     },
     lastName: {
-        type: String,
-        required: true
+      type: String,
+      required: true,
     },
     email: {
-        type: String,
-        required: true,
-        unique: [true, 'email already in database']
+      type: String,
+      required: true,
+      unique: [true, "email already in database"],
     },
     password: {
-        type: String,
-        required: true
+      type: String,
+      required: true,
     },
-    phone: Number,
-    state: {
-        type: String,
-        required: true
+    phoneNumber: {
+      type: Number,
+      required: true,
+      unique: true,
     },
-    category: {
-        type: String,
-        enum: ['individual', 'organisation'],
-        default: 'individual',
-        required: true
-    },
-    organizationName: String
-},
-    {
-        timestamps: true
-    }
-)
+  },
+  {
+    timestamps: true,
+  }
+);
 
 donorSchema.pre("save", function (next) {
-    let user = this;
-    bcrypt
-      .hash(user.password, 10)
-      .then((hash) => {
-        user.password = hash;
-        next();
-      })
-      .catch((error) => {
-        console.log(`Error in hashing password: ${error.message}`);
-        next(error);
-      });
-  });
-  
-  donorSchema.methods.passwordComparison = function (inputPassword) {
-    let user = this;
-    return bcrypt.compare(inputPassword, user.password);
-  };
-  
+  let user = this;
+  bcrypt
+    .hash(user.password, 10)
+    .then((hash) => {
+      user.password = hash;
+      next();
+    })
+    .catch((error) => {
+      console.log(`Error in hashing password: ${error.message}`);
+      next(error);
+    });
+});
+
+donorSchema.methods.passwordComparison = function (inputPassword) {
+  let user = this;
+  return bcrypt.compare(inputPassword, user.password);
+};
 
 //Creating a Donor Model with the above schema
-const donor = mongoose.model('Donor', donorSchema);
+const donor = mongoose.model("Donor", donorSchema);
 
 //Exporting the donor model
 module.exports = donor;
