+++ node_src/routes/mobile/donor/index.js
@@ -1,14 +1,14 @@
 const passport = require("passport");
 const secureRoute = require("./secured-routes");
-const undergraduateRoutes = require("./undergraduates");
-require("../undergraduates/auth");
+const donorRoutes = require("./donor");
+require("../donor/auth");
 
 const router = require("express").Router();
 
-router.use("/undergraduate", undergraduateRoutes);
+router.use("/donor", donorRoutes);
 // Plug in the JWT strategy as a middleware so only verified users can access this route.
 router.use(
-  "/undergraduate",
+  "/donor",
   passport.authenticate("jwt", { session: false }),
   secureRoute
 );
