+++ node_src/routes/mobile/fundme.js
@@ -1,0 +1,12 @@
+const router = require("express").Router();
+const beneficiary = require("./beneficiary");
+const donor = require("./donor");
+
+router.get("/", (req, res) => {
+  res.send("Welcome to the home page");
+});
+
+router.use("/", beneficiary);
+router.use("/", donor);
+
+module.exports = router;
